% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_7for_lefse.R
\name{data_7for_lefse}
\alias{data_7for_lefse}
\title{Step7 Prepare data for LEfSe analysis, the input data are from the result of postUsearch step1-6}
\usage{
data_7for_lefse(group, genus, taxa)
}
\arguments{
\item{group}{the grouping metadata}

\item{genus}{the ASVs abundance table, adapted from ASVs_norm.txt using sort_1ASVs function}

\item{taxa}{the taxa data adapted from result/ASV_tax2.txt using sort_2Taxa function}
}
\value{
a data frame for LEfSe ananlysis
}
\description{
Step7 Prepare data for LEfSe analysis, the input data are from the result of postUsearch step1-6
}
\examples{
data_7for_lefse(group = group, genus = genus, taxa = taxa)
}
\author{
Zhonghui Gai
}
