% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_slice.R
\name{group_slice}
\alias{group_slice}
\title{Step6 Select specific rows from the grouping information}
\usage{
group_slice(
  select = c("CTL$", "UC$", "LA$"),
  vector = group$group,
  data = NULL
)
}
\arguments{
\item{select}{the groups to be selected, use ^ or $ please.}

\item{vector}{the group vector to be select from}

\item{data}{the data frame to be subset}
}
\value{
a logical vector with the same long as the input vector or a data frame
}
\description{
Step6 Select specific rows from the grouping information
}
\examples{
ind <- group_slice(select = c("CTL$", "UC$", "LA$"), vector = group$group, data = sum.phylum)
}
\author{
Zhonghui Gai
}
