% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_1ASVs.R
\name{sort_1ASVs}
\alias{sort_1ASVs}
\title{Step1 Read the ASVs data and sort by copies number}
\usage{
sort_1ASVs(asv.path = "ASVs_norm.txt", percent = TRUE, trans = TRUE)
}
\arguments{
\item{asv.path}{The path of the ASVs_norm.txt data}

\item{percent}{Transform to abundance percent}

\item{trans}{a boolean value, Inverse the matrix}
}
\value{
a data frame containg the ASVs
}
\description{
Step1 Read the ASVs data and sort by copies number
}
\examples{
sort_1ASVs(asv.path = "result/ASVs_norm.txt")
}
\author{
Zhonghui Gai
}
