% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_2Taxa.R
\name{sort_2Taxa}
\alias{sort_2Taxa}
\title{Step2 Read the classification data of the ASVs}
\usage{
sort_2Taxa(taxa.path = "ASV_tax2.txt", taxa = NULL)
}
\arguments{
\item{taxa.path}{The path of taxa data, usually ASV_tax2.txt, this data also useful to lefse analysis}

\item{taxa}{a value to indicate whether the data already exist}
}
\value{
a data frame containing the classification information and the lineage
}
\description{
Step2 Read the classification data of the ASVs
}
\examples{
sort_2Taxa(taxa.path = "result/ASV_tax2.txt")
}
\author{
Zhonghui Gai
}
