% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_3group.R
\name{sort_3group}
\alias{sort_3group}
\title{Step3 Read the metadata containg the grouping information}
\usage{
sort_3group(group.path = "group.xlsx")
}
\arguments{
\item{group.path}{The path of the grouping metadata, which should be saved in xlsx format. The data contains two collumn, one for samples and another for grouping}
}
\value{
a data frame containg the grouping information, and also sorted by the order of rownames of genus
}
\description{
Step3 Read the metadata containg the grouping information
}
\examples{
sort_3group(group.path = "group.xlsx")
}
