% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_4alpha.R
\name{sort_4alpha}
\alias{sort_4alpha}
\title{Step4 Read in the alpha diversity infromation, which genarated by Usearch}
\usage{
sort_4alpha(alpha.path = "alpha/alpha.txt", extract = TRUE)
}
\arguments{
\item{alpha.path}{The path of the alpha data, usually alpha.txt}

\item{extract}{a boolean value to indicate whether extract the useful information, the default value is TRUE}
}
\value{
a data frame containg the alpha diversity information
}
\description{
Step4 Read in the alpha diversity infromation, which genarated by Usearch
}
\examples{
alpha <- sort_4alpha(alpha.path = "alpha/alpha.txt")

}
\author{
Zhonghui Gai
}
