% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_5sum_pfg.R
\name{sort_5sum_pfg}
\alias{sort_5sum_pfg}
\title{Step5 Read in the summary data from tax dictionary at the phylum, family and genus levels}
\usage{
sort_5sum_pfg(sum.path = "tax/sum_phylum.txt")
}
\arguments{
\item{sum.path}{the path of the sum data}
}
\value{
a data frame containning grouping information
}
\description{
Step5 Read in the summary data from tax dictionary at the phylum, family and genus levels
}
\examples{
sum.phylum <- sort_5sum_pfg(sum.path = "tax/sum_phylum.txt")
sum.family <- sort_5sum_pfg(sum.path = "tax/sum_family.txt")
sum.genus <- sort_5sum_pfg(sum.path = "tax/sum_genus.txt")
}
\author{
Zhonghui Gai
}
